+++ mobile/test/presentation/widgets/timeline/timeline_with_grouping_pill_test.dart
@@ -5,11 +5,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:immich_mobile/domain/models/asset/base_asset.model.dart';
+import 'package:immich_mobile/domain/models/settings_key.dart';
 import 'package:immich_mobile/domain/models/store.model.dart';
 import 'package:immich_mobile/domain/models/timeline.model.dart';
 import 'package:immich_mobile/domain/models/timeline_temporal_scope.model.dart';
 import 'package:immich_mobile/domain/models/user.model.dart';
-import 'package:immich_mobile/domain/models/settings_key.dart';
 import 'package:immich_mobile/domain/services/store.service.dart';
 import 'package:immich_mobile/domain/services/timeline.service.dart';
 import 'package:immich_mobile/domain/services/user.service.dart';
@@ -144,7 +144,7 @@ TimelineService _service() {
   return TimelineService((
     bucketSource: () => Stream.value([TimeBucket(date: DateTime(2025, 1), assetCount: 4)]),
     assetSource: (offset, count) async {
-      final end = (offset + count).clamp(0, assets.length).toInt();
+      final end = (offset + count).clamp(0, assets.length);
       if (offset >= end) return const <BaseAsset>[];
       return assets.sublist(offset, end);
     },
